headsets.json: add entry for Homido VR Headset
Note that this was created by taking the fully-unpacked URL from the unusual
QR code on http://www.homido.com/en/articles/configure-cardboard-apps-rendering-for-homido
and running it through goo.gl myself: the headset itself likely has a different
QR code (probably with a different shortened URL, possibly pointing to a
different profile). If I ever encounter it, that profile will supersede
this one.