Remove headset showcase
Ultimately, this is going to come down to either scanning the QR code
or selecting from a list, with the latter being the less-proritized
selection mechanism. If the application wants to present their own
opinionated menu of Cardboard devices, they can chose to do so and
ignore the user's selection from whatcardboard.