Remove implicit message for stored profile on iframe startup
Not only would this message make dealing with a reloaded iframe
after applying updates more complex, it's also being put forth at
a point where the application can't necessarily use the information
(for instance, because it only loads its VR mode when requested).
Better to only return the info in response to a `query` message,
and have consumers send that message if and when they want info
about the stored profile.